dlc_oracle_bot/requirements.txt
alembic
colorama
python-dotenv
PyMySQL
structlog
SQLAlchemy
